+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [2.0.7] 2017-02-28
 ### Added
 - Ability for integration test to check for attachment name